+++ packages/app/cypress/e2e/settings.cy.ts
@@ -166,7 +166,7 @@ describe('App: Settings', () => {
       cy.get(SidebarSettingsLinkSelector).click()
       cy.findByText('Project settings').click()
       cy.get('[data-cy="file-match-indicator"]').contains('19 matches')
-      cy.get('[data-cy="spec-pattern"]').contains('tests/**/*.(js|ts|coffee)')
+      cy.get('[data-cy="spec-pattern"]').contains('tests/**/*.(js|ts)')
     })
 
     it('shows the Experiments section', () => {
@@ -249,7 +249,7 @@ describe('App: Settings', () => {
 
       cy.get('[data-cy="config-code"]').within(() => {
         cy.get('[data-cy-config="config"]').contains('tests/_fixtures')
-        cy.get('[data-cy-config="config"]').contains('tests/**/*.(js|ts|coffee)')
+        cy.get('[data-cy-config="config"]').contains('tests/**/*.(js|ts)')
         cy.get('[data-cy-config="config"]').contains('tests/_support/spec_helper.js')
         cy.get('[data-cy-config="env"]').contains('REMOTE_DEBUGGING_PORT')
         cy.get('[data-cy-config="env"]').contains('INTERNAL_E2E_TESTING_SELF')